requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics==8.0.202
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pydantic==2.4.2

# Тестовые зависимости
pytest==7.4.3
httpx==0.25.2
